[Contract API] Use promise to handle multiple contact deploy
