+++ routes/ethereum.js
@@ -7,6 +7,7 @@ var contract = require('../models/contract.js');
 var contractFunction = require('../models/contractFunction.js');
 var contractEvent = require('../models/contractEvent.js');
 var ethereumController = require('../controllers/ethereumController.js');
+var contractController = require('../controllers/contractController.js');
 var sqsHelper = require('../helpers/aws/sqsHelper.js');
 var web3 = new Web3();
 var router = express.Router();
@@ -58,65 +59,9 @@ router.get('/v1/contract', function (req, res, next) {
 router.post('/v1/contract', function (req, res, next) {
     if (req.body.sourceCode !== undefined && req.body.sourceCode !== null && req.body.sourceCode !== '') {
     	var sourceCode = req.body.sourceCode;
-        var result = solc.compile(sourceCode, 1);
-        var id = [];
-        for (var contractName in result.contracts) {
-        	var abi = result.contracts[contractName].interface;
-            var contractEntity = {
-                name: contractName, 
-                sourceCode: sourceCode,
-                byteCode: result.contracts[contractName].bytecode,
-                language: result.contracts[contractName].metadata.language,
-                compilerVersion: result.contracts[contractName].metadata.compiler,
-                abi: abi,
-                gasEstimates: web3.eth.estimateGas({data: '0x' + result.contracts[contractName].bytecode})
-            };
-
-            contract.create(contractEntity).then(function (contractId) {
-            	console.log('[CONTRACT CREATE] id: ' + contractId);
-                id.push(contractId);
-                
-                var message = {
-                	"contractId": contractId
-                }
-                sqsHelper.send(JSON.stringify(message), 
-                	process.env.AWS_CONTRACT_QUEUE_URL, 10, 
-                	'contract');
-                JSON.parse(abi).forEach(function(data){
-                    if (data.type === 'event') {
-                        var contractEventEntity = {
-                            contractId: contractId,
-                            eventName: data.name,
-                            eventParameters: data
-                        };
-                        // TODO insert contractEvent and contractFunction
-                        contractEvent.create(contractEventEntity).then(function (contractEventId) {
-                            console.log('[CONTRACTEVENT CREATE] id: ' + contractEventId);
-                        }).catch(function (err) {
-                            console.log(err.message, err.stack);
-                        });
-                    }
-                    if (data.type === 'function') {
-                        var contractFunctionEntity = {
-                            contractId: contractId,
-                            functionName: data.name,
-                            functionParameters: data
-                        };
-                        contractFunction.create(contractFunctionEntity).then(function (contractFunctionId) {
-                            console.log('[CONTRACTFUNCTION CREATE] id: ' + contractFunctionId);
-                        }).catch(function (err) {
-                            console.log(err.message, err.stack);
-                        });
-                    }
-                });
-            }).catch(function (err) {
-                // error handle
-                console.log(err.message, err.stack);
-                res.json({'error': {'message': err.message}});
-            });
-        }
-        // TODO add Promise
-        res.json({'data': {'contractId': id}});
+        contractController.deployContract(sourceCode).then(function(ids){
+        	res.json({'data': {'contractId': ids}});
+        })
     } else {
         console.log('error invalid source code!');
         res.json({'error': {'message': 'invalid source code'}});
