+++ tests/models/eventDataTest.js
@@ -1,0 +1,7 @@
+var eventData = require('../../models/eventData.js');
+
+eventData.readAll().then(function(eventDataResult) {
+	console.log(eventDataResult.rowCount);
+}).catch(function(err) {
+	console.log(err);
+});
